Use FLOAT instead of INT to store metrics
git-svn-id: https://svn.codehaus.org/sonar-plugins/trunk@802 d25a0af3-c461-0410-a42d-853a01480fe4

+++ quality-index/src/main/java/org/sonar/plugins/qi/QIMetrics.java
@@ -9,19 +9,19 @@ import java.util.Arrays;
 public class QIMetrics implements org.sonar.api.measures.Metrics {
 
   public static final Metric QI_QUALITY_INDEX = new Metric("qi-quality-index", "Quality Index",
-    "The quality index of a project", Metric.ValueType.INT, 1, true, CoreMetrics.DOMAIN_GENERAL);
+    "The quality index of a project", Metric.ValueType.FLOAT, 1, true, CoreMetrics.DOMAIN_GENERAL);
 
   public static final Metric QI_CODING_VIOLATIONS = new Metric("qi-coding-violations", "Coding Violations",
-    "Coding Violations", Metric.ValueType.INT, -1, true, CoreMetrics.DOMAIN_RULES);
+    "Coding Violations", Metric.ValueType.FLOAT, -1, true, CoreMetrics.DOMAIN_RULES);
 
   public static final Metric QI_STYLE_VIOLATIONS = new Metric("qi-style-violations", "Style Violations",
-    "Style Violations", Metric.ValueType.INT, -1, true, CoreMetrics.DOMAIN_RULES);
+    "Style Violations", Metric.ValueType.FLOAT, -1, true, CoreMetrics.DOMAIN_RULES);
 
   public static final Metric QI_COMPLEXITY = new Metric("qi-complexity", "Complexity",
-    "Complexityt", Metric.ValueType.INT, -1, true, CoreMetrics.DOMAIN_COMPLEXITY);
+    "Complexityt", Metric.ValueType.FLOAT, -1, true, CoreMetrics.DOMAIN_COMPLEXITY);
 
   public static final Metric QI_TEST_COVERAGE = new Metric("qi-test-coverage", "Test Coverage",
-    "Test Coverage", Metric.ValueType.INT, -1, true, CoreMetrics.DOMAIN_TESTS);
+    "Test Coverage", Metric.ValueType.FLOAT, -1, true, CoreMetrics.DOMAIN_TESTS);
 
   public static final Metric QI_COMPLEXITY_FACTOR = new Metric("qi-complexity-factor", "Complexity Factor",
     "Complexity Factor", Metric.ValueType.PERCENT, -1, true, CoreMetrics.DOMAIN_COMPLEXITY);
